Refactor for contao/phpstan namespace
